Issue #4106: split and organized Checkstyle inputs by Test for AbstractClassName
